Add files via upload
This folder won't be here very long. This is not a part of the program for anyone to contribute to. This is a test thing that I built for the SQLite database, so that I would learn how to do a few things, which I need to implement in GLIDE. This will be gone, as soon as I have what I need implemented into GLIDE.