Add files via upload
This file has a the calculations that GLIDE needs to use for scheduling repetitions, and for calculating the forgetting curve. These are based on the research document that is linked to the dartomic web site, under the "LATEST" section. I hope no one looks at this yet, because I did not originally intend on uploading it for others to see like this. It's kind of a mess. It will look better once I implement what I need out of it, into GLIDE, and then this file will be gone.